refactor(View): 자동차의 이름을 요청 후 검증 로직을 이동
InputView에서 구현한 자동차의 이름을 입력받고 검증하는 로직을 View 클래스로 이동한다.
