+++ src/main/java/racingcar/view/InputView.java
@@ -26,29 +26,5 @@ public final class InputView {
         }
     }
 
-    private static void validateInvalidDelimiter(String names) {
-        if (isEdgeDelimiter(names) || hasContinuousDelimiters(names)) {
-            throw RaceException.of(ErrorMessage.INVALID_DELIMITER_ERROR);
-        }
-    }
 
-    private static boolean hasContinuousDelimiters(String names) {
-        return names.contains(generateContinuousDelimiters());
-    }
-
-    private static String generateContinuousDelimiters() {
-        return DELIMITER.getSymbol().repeat(2);
-    }
-
-    private static boolean isEdgeDelimiter(String names) {
-        return startWithDelimiter(names) || endsWithDelimiter(names);
-    }
-
-    private static boolean startWithDelimiter(String names) {
-        return names.startsWith(DELIMITER.getSymbol());
-    }
-
-    private static boolean endsWithDelimiter(String names) {
-        return names.endsWith(DELIMITER.getSymbol());
-    }
 }
